changed startup sequence of events
for the startup animations option to work, the startup sequence creates the gpaths for the digits earlier in the process and places their layers on the screen first thing, in the case where startup animations are disabled.  otherwise, they start off screen as before.  

also created a new function to handle the positioning of the origin point for the time digits.